Remove ancient use of ldd
Once upon a time, there was no go-jsonnet, and `kubecfg` used cgo to link against a static (C++) libjsonnet.

When the dinosaurs downloaded this ancient pre-compiled kubecfg binary, it was important that it _was_ statically linked, and did not contain references to external shared libraries that those dinosaurs might not have had locally.   There was a line in the ancient build rituals to help detect and debug such issues.

Remove this long-obsolete reference to the ancient rituals, and let them pass into git history.